backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.1
python-docx==1.1.2
rank-bm25==0.2.2
requests==2.32.3
pydantic==2.9.0
